Domain Management: Transfer: Move transfer out stuff  to ./transfer/transfer-out
