+++ assets/stylesheets/_components.scss
@@ -348,7 +348,7 @@
 @import 'my-sites/upgrades/checkout-thank-you/style';
 @import 'my-sites/upgrades/checkout-thank-you/google-voucher/style';
 @import 'my-sites/upgrades/components/domain-warnings/style';
-@import 'my-sites/upgrades/domain-management/transfer/style';
+@import 'my-sites/upgrades/domain-management/transfer/transfer-out/style';
 @import 'my-sites/resume-editing/style';
 @import 'my-sites/stats/stats-list/style';
 @import 'notices/style';
