+++ client/my-sites/upgrades/domain-management/transfer/transfer-out/transfer-prohibited.jsx
